BL-1815 Reorganize helper methods into smaller files
