+++ app/helpers/application_helper.rb
@@ -1,15 +1,6 @@
 # frozen_string_literal: true
 
 module ApplicationHelper
-  # value ex: "MAIN stacks"
-  def render_location(value)
-    params = value.to_s.split
-    [ Rails.configuration.libraries[params.first],
-      Rails.configuration.locations.dig(*params) ]
-      .compact
-      .join(" ")
-  end
-
   def aeon_request_url(document)
     form_fields = {
          ItemTitle: document.fetch("title_statement_display", ""),
@@ -45,62 +36,6 @@ module ApplicationHelper
 
     if libraries.include?("SCRC")
       button_to(t("requests.aeon_button_text"), aeon_request_url(document), class: "btn btn-primary")
-    end
-  end
-
-  def total_items(results)
-    results.total_items[:query_total] || 0 rescue 0
-  end
-
-  def total_online(results)
-    results.total_items[:online_total] || 0 rescue 0
-  end
-
-  def bento_engine_nice_name(engine_id)
-    I18n.t("bento.#{engine_id}.nice_name")
-  end
-
-  def bento_icons(engine_id)
-    case engine_id
-    when "articles"
-      content_tag(:span, "", class: "bento-icon bento-article m-3")
-    when "journals"
-      content_tag(:span, "", class: "bento-icon bento-journal m-3")
-    when "databases"
-      content_tag(:span, "", class: "bento-icon bento-database m-3")
-    when "books_and_media"
-      content_tag(:span, "", class: "bento-icon bento-book m-3")
-    when "library_website"
-      content_tag(:span, "", class: "bento-icon bento-website m-3")
-    when "lib_guides"
-      content_tag(:span, "", class: "bento-icon bento-guide m-3")
-    end
-  end
-
-  def bento_link_to_full_results(results)
-    total = number_with_delimiter(total_items results)
-    BentoSearch.get_engine(results.engine_id).view_link(total, self)
-  end
-
-  # TODO: move to decorator or engine class.
-  def bento_link_to_online_results(results)
-    total = number_with_delimiter(total_online results)
-
-    case results.engine_id
-    when "blacklight", "books_and_media"
-      url = search_catalog_path(q: params[:q], f: { availability_facet: ["Online"] })
-      link_to "View all #{total} online items", url
-    when "journals"
-      url = search_journals_path(q: params[:q], f: { availability_facet: ["Online"] })
-      link_to "View all #{total} online journals", url
-    when "articles"
-      url = url_for(
-        action: :index, controller: :primo_central,
-        q: params[:q], f: { availability_facet: ["Online"] }
-      )
-      link_to "View all #{total} online articles", url
-    else
-      ""
     end
   end
 
@@ -216,84 +151,6 @@ module ApplicationHelper
     if !link.blank?
       link_to(t("blacklight.banner_link"), link)
     end
-  end
-
-  def query_list(title, tooltip, query, footer_field = nil)
-    if @document&.id
-      query += "&filter_id=#{@document.id}"
-    end
-
-    if footer_field
-      query += "&footer_field=#{footer_field}"
-    end
-
-    title = link_to title, search_catalog_path + "?#{query}"
-
-    render partial: "query_list/results", locals: { query: query + "&per_page=5", title: title, tooltip: tooltip }
-  end
-
-  def query_list_footer_value(document, field)
-    if document[field]&.include?(0)
-      document[field].delete(0)
-    end
-
-    value = document[field]&.first
-
-    case field
-    when "date_added_facet"
-      begin
-        Date.parse("#{value}").strftime("%Y-%m-%d")
-      rescue => e
-        Honeybadger.notify("Error trying to parse date_added_facet value; @htomren " + "#{e.message}")
-        ""
-      end
-    else
-      value
-    end
-  end
-
-  def creator_query_list(document)
-    field = "creator_display"
-    creator = document[field]&.first&.split("|")&.first
-    query_list(
-      title = "By author/creator",
-      tooltip = t("tooltip.creator"),
-      query = "f[creator_facet][]=#{creator}&sort=pub_date_sort+desc,+title_sort+asc", footer_field = "pub_date"
-    ) unless creator.blank?
-  end
-
-  def call_number_query_list(document, order = "asc")
-    field = "lc_call_number_display"
-    lc_call_number = document["lc_call_number_display"]&.first&.gsub(/\s+/, "")
-    # This is built this way due to a possible bug with desc lc_call_number sorts
-    # We may need to refactor this when that bug is fixed
-    # We think the bug is in the add_lc_range_search_to_solr method
-    position = order == "asc" ? "begin" : "end"
-    call_number_title = order == "asc" ? "By call number (a-z)" : "By call number (z-a)"
-    tooltip = order == "asc" ? t("tooltip.call_number_asc") : t("tooltip.call_number_desc")
-    query_list(
-      title = call_number_title,
-      tooltip = tooltip,
-      query = "f_1=all_fields&f_2=all_fields&f_3=all_fields&op_1=AND&op_2=AND&operator%5Bq_1%5D=contains&operator%5Bq_2%5D=contains&operator%5Bq_3%5D=contains&q_1=&q_2=&q_3=&range%5Blc_classification%5D%5B#{position}%5D=#{lc_call_number}&range%5Bpub_date_sort%5D%5Bbegin%5D=&range%5Bpub_date_sort%5D%5Bend%5D=&search_field=advanced&sort=lc_call_number_sort+#{order}%2C+pub_date_sort+desc", footer_field = field
-    ) unless lc_call_number.blank?
-  end
-
-  def libraries_query_display(document)
-    libraries = document["library_facet"]
-    unless libraries.nil?
-      if libraries.count > 2
-        html = content_tag(:p, libraries.first, class: "mb-0 pb-0 text-truncate")
-        html += content_tag(:p, "More Locations", class: "mb-0 pb-0 font-italic")
-      elsif libraries.count === 2
-        content_tag(:p, libraries.join("<br />").html_safe, class: "mb-0 pb-0 text-truncate")
-      else
-        content_tag(:p, libraries.join("<br />").html_safe, class: "mb-0 pb-0")
-      end
-    end
-  end
-
-  def query_list_view_more_links(query_params)
-    link_to "View More", search_catalog_path(query_params.except(:per_page)), class: "query-list-view-more stretched-link"
   end
 
   def format_classes_for_icons(document)
